[Refactor] Make COLORS auto-repeat with input
The COLORS array was created by manually and repeatedly putting in the
same colors.

This is refactored to requiring only putting in a list of unique colors,
which will then use code to automatically repeat once.

The 'shuffledColors' variable is also now declared using 'const', since
it does not get re-assigned.