[Feat] Finish minimum viable product (card game)
Added the handling function for when cards are clicked, the
storing/tracking of pairs of cards, the comparing a pair of cards, and
the win condition.

The color property for the card elements were placed into a custom
attribute for each card to more easily access it.  There was no need to
put it in the class attribute, along with the other classes.

In the handleCardClick function, checks for if a card has the
"revealed" class and if the second card has been chosen were used,
because it was a simple way to not count the same card as a match when
clicking twice and of pausing the game after clicking on two cards.