[Feat] Implement score tracking
Added displaying of the current game score and the best score made
across games, on the HTML page.  Added the updating of the score after
clicking on a card and the storing of the best score after winning the
game.

These were added in response to the extra features to include for the
project.