[Feat] Add start and restart
Added a "Start" button when loading the HTML and a "Restart" button
after completing a game.  This was in response to extra features that
can be placed into the project.

The displaying of the best score was moved into its own function, so
that it can be easily called whenever a new game starts.

The shuffling of cards was placed into a startGame function, which made
it possible to have a new random order of cards for every game restart.

The updateScore function was changed to setScore, which took in an
argument for the score, allowing the score to be more easily reset for
every game restart.