[Feat] Add CSS styling
Moved the start and restart buttons to be between the score and best
score to be consistent with the placement of info and because it was
easier to work with it there.

Moved the score and best score text into spans, and also into score
area divs to more easily control positioning in CSS.

Created a removeButton function since buttons in the new location do
not conveniently get removed.  Since there should only be one button at
a time in the button area, finding the first element child of the
button area will work fine.